Return response as error if not valid json object
If the response returned by Mistral do not contain a valid json an
exception is thrown by the client due to invalid json format.
If there's a ValueError exception and it's an error code try to return
the response text instead. (Ex: Authorization Failed )

This may be a temporary fix if the BE is able to return the json
error msg returned by keystone.

Change-Id: I540f3aaa35fb15adc04bf0a33e46af8ac614c814
Closes-Bug: #1316298