Change the mistralclient for Mistral action pack
The purpose of this patch is to make the new Mistral action pack
actions more useful when they are coded in a workflow or workbook.

The new Mistral action pack actions are executed by calling the
appropriate python-mistralclient API methods. The API methods to
create/update/validate the workbook/workflow/action/environment
resources do not read the definitions/environments from files as the
CLI commands do. They require the actual definition/environement be
passed in as an argument. This can lead to problems when trying to
use the Mistral action pack actions from within a workflow or
workbook.

For example, the workflow create/update/validate methods all take a
definition argument. The mistralclient APIs are currently coded to
take the actual definition, and not the definition file name as the
CLI does. This leads to problems when coding the
mistral.workflows_create action in a workflow for example. If the
actual definition is passed into the action in the workflow YAML, any
YAQL expressions in the workflow being created by that action get
evaluated when read, and are not stored in the database as desired.
This essentially makes it useless unless the workflow you are
creating does not contain any YAQL. That is unrealistic.

This patch enhances those methods of the mistralclient so that they
can take either the actual definition, or the definition file name or
file URI. This allows the action to be coded in a workflow in a way
similar to the CLI by allowing the definition file name, or file URI
to be specified instead of the actual definition. The change allows
either the actual definition or the definition file name, or file
URI to be specified when calling the affected mistralclient API
methods. No changes to the API signatures are required. Each method
affected will determine if the value passed in is a definition file
name or file URI, or the actual definition.

New unit test cases were also added for these changes.

Change-Id: I9ca07bedf685839db130693a7aa6e091a9b8c4ad
Implements: blueprint mistral-mistral-actions