Make osprofiler dependency "soft"
All OpenStack clients contain osprofiler as a soft dependency
(profiling works only if osprofiler package is installed already
on the env). This is done intentionally: this allows to avoid its
installation on the client nodes, that are not supposed to present
a profiling client (e.g. non-admin users). Moreover, marking osprofiler
as a strict dependency now leads to bug #1361235 being reproduced
in specific Horizon installations (when horizon is installed in a
python virtualenv along with it's dependencies).

Change-Id: Ifdfaa89995fb2340d7b6540373f41e9f2fd35bcc
Closes-Bug: #1361235