remove openstack/common/cliutils.py
As part of the first community-wide goal [1], teams were asked
to remove the openstack/common package of their projects
if one existed. This was a byproduct of the old oslo-incubator
form of syncing common functionality.

The module removed in this patch, cliutils, was removed, and the
function env() was moved to shell.py since that was the only
reference to it.

Additional patches will continue to trim content from openstack/
common

[1] http://governance.openstack.org/goals/ocata/remove-incubated-oslo-code.html

Change-Id: I0d87750e6485f34272e69182914cd2659e9439b4