Add --utc flag to cron trigger create
The server should not be doing any translation of times. It should be
assumed that what is sent to the server is being sent as UTC. The
addition of the --utc flags facilitates this. If the --utc flag is
specified, then the --first-time passed is assumed to be already
UTC and will not undergo a conversion. If it is not passed, the
--first-time arg will be converted based on the the localtime of the
client, then sent as UTC.

Change-Id: I424d042d1cbdadae3aa14cd408d9f27d14823aef
Depends-On: Ifbd63d9e1f56085928bede22ce4f2954e1b38991
Closes-Bug: #1654218