Set the default value of --limit parameter
The "--limit" parameter of heavy CLI commands should not be
infinite by default. For commands like "mistral task-list" and
"execution-list" we need to set --limit parameter by default to
some reasonable value. The proposal is 100. Otherwise it's easy
to cause a huge load on a server if the result set is not limited.

A warning message is thrown whenever these commands are executed
without using "--limit" parameter.

Change-Id: If0fead181171e5b9c590f3f4c852c6ad2541ff94
Implements: blueprint mistral-cli-default-limit-parameter