mistral execution-list -f value should be empty
The '-f value' option in the CLI is intended to be used by scripting
etc. as it gives you the raw value. This can make life easier in bash.
However, if there are no results mistral will return <none> - this isn't
the raw value, and means you need to map it via case statement.

Related-Bug: 1691435

Change-Id: I246f772253e0fdb4dc8a82a4d4186b4bb6626981