Support for Project_domain_id and user_domain_id.
Currently in Mistral-client There is support for only
PROJECT_DOMAIN_NAME but in the latest release of OpenStack,
there is an option PROJECT_DOMAIN_ID.This DOMAIN_ID option
is evident in few cloud environments and not evident in other
so it is better to give support for both DOMAIN_ID and DOMAIN_NAME.

In Mistral Client though the option to use target_user_domain_name
and target_project_domain_name are not used in the authentication.
This commit is to use them.
Added unit-testcases for the same and added few more for robustness.

Closes-Bug: #1712545
Closes-Bug: #1712739

Change-Id: I64aa2ff3324ba770c1e3aabb6410cca91c793c15