Don't create client for help and bash completion
Currently when we need a help, client object is created
and authentication is performed. This is completely useless
and leads to unnecessary actions in the background.

This patch:
1. Prevents creation of client object (and therefore
authentication) for help or bash-completion commands.
2. Removes a workaround from keystone auth module that disables
sending requests to the server if help or bash-completion
commands are executing.
3. Adds related unit tests.

Change-Id: Ia26d7f4e56f5ef3ae0ac5e94e8e77d1a78f8829e
Closes-bug: #1720795
(cherry picked from commit d9da161c16bc524d4e1d64f09fa2d857cbcf4537)