Create client for interactive shell
The client allows you to run an interactive mode that doesn't exit
between calls and reuses the client which is handy for end users.
Currently the code doesn't actually create a workflow engine client in
the interactive mode so all the mistral actions fail. We should create a
client if the client doesn't pass any args to the shell because that
enters interactive mode.

Change-Id: I8b4c56277f0f172da712f777c4c856dba6fa0b0b
Closes-Bug: #1861357