fixed compatibility issues between mistral and client
removed the namespace field from the action formatter to fix
compatibility issues between the client and older versions
of mistral.

now the action-list command takes the namespace
(with --namespace) as an argument and returns a list of the
actions in that namespace.

the command without --namespace would return the actions
of the default namespace ''.

fixed issues with execution-list command, the field "nulls"
is removed if its empty

Change-Id: Ibf64ff8d841ea1192b410902589b0c690b0b2fca
Signed-off-by: ali <ali.abdelal@nokia.com>